Add diagnostics tests
