+++ Parser/SyntaxNode.cs
@@ -126,38 +126,6 @@ namespace Parser
             }
         }
     }
-
-    public class SyntaxDiagnostic
-    {
-        public int Position { get; }
-
-        public static SyntaxDiagnostic From(Internal.TokenDiagnostic diagnostic, int Position)
-        {
-            switch (diagnostic)
-            {
-                case Internal.MissingTokenDiagnostic missingToken:
-                    return new MissingTokenSyntaxDiagnostic(Position, missingToken.Kind);
-            }
-
-            throw new System.ArgumentOutOfRangeException(nameof(diagnostic));
-        }
-
-        protected SyntaxDiagnostic(int position)
-        {
-            Position = position;
-        }
-    }
-
-    public class MissingTokenSyntaxDiagnostic : SyntaxDiagnostic
-    {
-        public TokenKind Kind { get; }
-
-        internal MissingTokenSyntaxDiagnostic(int position, TokenKind tokenKind)
-            : base(position)
-        {
-            Kind = tokenKind;
-        }
-    }
     
     public abstract class StatementSyntaxNode : SyntaxNode
     {
